backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0
python-dotenv>=1.0.0
slowapi>=0.1.9
faster-whisper>=1.0.3
python-multipart>=0.0.9
websockets>=12.0
silero-vad
torch
torchaudio
